TPose editor is functional and sort of working; it will try to keep animations unchanged by applying the inverse transform to keyframes of edited nodes. Still buggy and requires at least one keyframe in all animations one wishes to keep unchanged. Made model loading not run on the gui thread and added a small non-modal popup that informs the user that a model is being loaded Added RoundKeyframesPanel for rounding and clamping(towards 0) animation values "Simplify Keyframes of Selected Nodes" and "Simplify Keyframes" has been merged into "Simplify Keyframes" and will now always bring up the SimplifyKeyframesPanel All Particle classes now extends EmitterIdObject

+++ modelstudio/src/com/hiveworkshop/rms/editor/model/util/ModelSaving/GeosetToMdlx.java
@@ -133,10 +133,9 @@ public class GeosetToMdlx {
 				short[] skinBoneWeights = vertex.getSkinBoneWeights();
 				Bone[] skinBoneBones = vertex.getSkinBoneBones();
 				for (int j = 0; j < 4; j++) {
-//					int index = skinBoneBones[j] == null ? 0 : skinBoneBones[j].getObjectId(model);
 					int index = skinBoneBones[j] == null ? 0 : model.getObjectId(skinBoneBones[j]);
 					mdlxGeoset.skin[(i * 8) + j] = (short) index;
-					mdlxGeoset.skin[(i * 8) + j + 4] = (short) (skinBoneWeights[j]);
+					mdlxGeoset.skin[(i * 8) + j + 4] = skinBoneWeights[j];
 					mdlxGeoset.tangents[(i * 4) + j] = vertex.getTangent().toFloatArray()[j];
 				}
 			}
